move top code to separate file
